відредаговані поля деякіх класів, добавлено функції наповнення списків

+++ spd1/Goods.cs
@@ -9,6 +9,6 @@ namespace spd1
     class Goods
     {
         public string Name;
-        public int Time;
+        public long Time;
     }
 }
