+++ spd1/Order.cs
@@ -9,6 +9,6 @@ namespace spd1
     class Order
     {
         public string Name;
-        public int TimeLeft;
+        public long TimeLeft;
     }
 }
